代码暂存

+++ lib/store/user_controller.dart
@@ -1,7 +1,13 @@
 
 
 import 'package:get/get.dart';
+import 'package:logger/logger.dart';
 import 'package:op_flutter/models/users/user_model.dart';
+import 'package:op_flutter/network/login/api.dart';
+import 'package:op_flutter/network/login/login_request.dart';
+import 'package:op_flutter/routes/app_routes.dart';
+import 'package:op_flutter/utils/common.dart';
+import 'package:op_flutter/widgets/custom_loading_dialog.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:convert';
 
@@ -32,18 +38,33 @@ class UserController extends GetxController {
     return User.fromJson(jsonDecode(json));
   }
 
+  /// 登出
   Future<void> logout() async {
-    final prefs = await SharedPreferences.getInstance();
+    try {
+      if (user.value != null) {
+        withLoadingDialog(() async {
+          // 清理本地缓存，保留 tid/uid
+          final prefs = await SharedPreferences.getInstance();
+          String? tid = prefs.getString('terminal');
+          String? uid = prefs.getString('uid');
 
-    // ⚠️ TID & UID 保留，其余需要清空
-    String? tid = prefs.getString('terminal');
-    String? uid = prefs.getString('uid');
+          await prefs.clear();
 
-    await prefs.clear();
+          if (tid != null) {
+            prefs.setString('terminal', tid);
+          }
+          if (uid != null) {
+            prefs.setString('uid', uid);
+          }
 
-    if (tid != null) prefs.setString('terminal', tid);
-    if (uid != null) prefs.setString('uid', uid);
-
-    user.value = null;
+          user.value = null;
+          Get.offAllNamed(AppRoutes.home);
+        });
+      }
+    } catch (e) {
+      // 可以记录错误或提示用户
+      final logger = Logger();
+      logger.e(e);
+    }
   }
 }
